Fold OPAM2 refinements into existing molgpka/pka_plugins
Addresses PR review feedback: reuses existing code instead of creating
parallel OPAM2 modules. Back-compatible with the stock MolGpKa flow —
plain MolGPKA() and load_model() callers are unaffected.

molgpka.py (additive):
- @lru_cache-backed load_model with torch.load(weights_only=True),
  falling back for older PyTorch.
- prepare_mol_for_prediction preserves heavy-atom indices.
- cap_unstable_sites bounds combinatorial protonation enumeration.
- opam_protonate_mol accepts SMILES or InChI and selectable output
  format, reusing modify_mol/get_pKa_data/modify_stable_pka/etc.
- MODELSEED_{ACID,BASE}_MODEL_PATH + MAX_UNSTABLE_SITES constants.

pka_plugins.py:
- Adds MolGPKAOpam2(MolGPKA) — inherits from MolGPKA, transitively from
  BasePkaPredictor, so the ChemAxon drop-in contract is preserved.
  Defaults to ModelSeed weights when present, supports InChI input,
  configurable output format, and caps combinatorial enumeration.

Removes the now-redundant src/molgpka_opam2.py and
src/pka_plugins_opam2.py.
